"For Volunteers
1. Waste Zones Identification
2. Heat Map Creation"

OK, so what I want to do exactly is it is for volunteering purpose. So what I want to do is that first of all, a form should open up which should list. I don't know. It should. It should do what I mean. It should open up a form where you ask the details of the zone that is dirty. I mean, the dirty place or anything, something like that. And then after using maps, you can select those place that is dirty, or use the gps location, or, I mean, tag the place where which is dirty. And then after. you submit it for listing. And then what it's gonna do is that there will be a map in the bottom navigation bar, which people can open and find. What are the dirty spots? You can even point them with a red icons or something like that. And that will be. the doty spots and what we can do more is that, yeah, listing. It's just the dirty places listing and viewing it in the map. First of all, the map should be added in the I mean in the center of the bottom navigation bar And in that map only there should be some icon or, I don't know, there should be an icon or a button to add listing, and that listing is gonna open up the form and something like that. I hope you understood.