"The map page now has a title and a "Report New Dirty Spot" button (currently non-functional). It displays 10 sample dirty spots in Kathmandu as red markers on the Google Map. Clicking a marker opens an InfoWindow with the spot's name. Below the map, there's a list of these dirty spots. Clicking an item in the list will also open the corresponding InfoWindow on the map."
I am using a development only application. So right now it's kinda like that. The billing is also not enabled for now. And yeah, the things that you have said is working. It's not working, I guess. So in the map stab, I don't know where you have added it. It's not working. It's not even shows that the title, the title is also not even shown. I don't know why. And I want you to make the map bigger, like take most of the space. I feel like that's gonna be better.